requirements.txt
aniso8601==4.1.0
atomicwrites==1.3.0
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
codecov==2.0.15
colorama==0.4.1
contextlib2==0.5.5
coverage==4.5.2
Flask==1.0.2
Flask-Cors==3.0.7
flask-restplus==0.12.1
Flask-Script==2.0.6
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
more-itertools==6.0.0
mysql-connector==2.2.9
nose==1.3.7
pluggy==0.8.1
py==1.8.0
PyJWT==1.7.1
pytest==4.3.0
python-dotenv==0.10.1
python-http-client==3.1.0
pytz==2018.9
requests==2.21.0
schema==0.7.0
sendgrid==6.0.5
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
